% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-funcs.R
\name{FT_from_height}
\alias{FT_from_height}
\title{Get Flight Time from Height}
\usage{
FT_from_height(height, gravity_const = 9.81)
}
\arguments{
\item{height}{Numeric vector in meters}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector. Flight time in seconds
}
\description{
\code{FT_from_height} returns flight time using \code{height}
}
\examples{
FT_from_height(0.3)
}
