% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-funcs.R
\name{FT_from_TOV}
\alias{FT_from_TOV}
\title{Get Flight Time from Take-off Velocity}
\usage{
FT_from_TOV(take_off_velocity, gravity_const = 9.81)
}
\arguments{
\item{take_off_velocity}{Numeric vector in m/s}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector. Flight time in seconds
}
\description{
\code{FT_from_TOV} returns flight time using \code{take_off_velocity}
}
\examples{
height_from_FT(0.7)
}
