% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-funcs.R
\name{height_from_FT}
\alias{height_from_FT}
\title{Get Height from Flight Time}
\usage{
height_from_FT(flight_time, gravity_const = 9.81)
}
\arguments{
\item{flight_time}{Numeric vector. Flight time in seconds}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector. Jump height in meters
}
\description{
\code{height_from_FT} returns jump height using \code{flight_time}
}
\examples{
height_from_FT(0.7)
}
