% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-funcs.R
\name{height_from_TOV}
\alias{height_from_TOV}
\title{Get Height from TOV}
\usage{
height_from_TOV(take_off_velocity, gravity_const = 9.81)
}
\arguments{
\item{take_off_velocity}{Numeric vector in m/s}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector
}
\description{
Function \code{height_from_TOV} calculates height reached using the ballistic equation
}
\examples{
height_from_TOV(2.75)
}
