% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate}
\alias{integrate}
\title{Trapezoidal integration over range of x}
\usage{
integrate(x, y, cumulative = FALSE, reverse = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{cumulative}{Should the cumulative vector be returned? Default is \code{FALSE}}

\item{reverse}{Should integration start from the end? Default is \code{FALSE}}
}
\value{
Numeric vector
}
\description{
Trapezoidal integration over range of x
}
\examples{
df <- dplyr::tibble(
  x = seq(0, 2 * pi, length = 100),
  y = sin(x),
  j = integrate(x, y, cumulative = TRUE)
)

df \%>\%
  ggplot2::ggplot(ggplot2::aes(x = x)) +
  ggplot2::geom_point(ggplot2::aes(y = y)) +
  ggplot2::geom_point(ggplot2::aes(y = j), color = "red")

# Total surface
integrate(df$x, df$y)

}
