% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longest_streak.R
\name{longest_TRUE_streak}
\alias{longest_TRUE_streak}
\title{Longest \code{TRUE} streak}
\usage{
longest_TRUE_streak(x)
}
\arguments{
\item{x}{A logical vector}
}
\value{
A numeric two-element vector with \code{start} and \code{stop} indices of the
longest \code{TRUE} value streak in \code{x}
}
\description{
Finds the location (i.e., start and stop index) of the longest \code{TRUE} streak
in the logical \code{x} vector
}
\examples{
longest_TRUE_streak(c(FALSE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE))

longest_TRUE_streak(1:10 > 3)

x <- seq(1, pi, length.out = 100)
y <- sin(x)

x[longest_TRUE_streak(y > 0.25 & y < 0.7)]

}
