% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-funcs.R
\name{TOV_from_force}
\alias{TOV_from_force}
\title{Get Take-off Velocity}
\usage{
TOV_from_force(force, mass, push_off_distance = 0.4, gravity_const = 9.81)
}
\arguments{
\item{force}{Numeric vector in Newtons}

\item{mass}{Numeric vector in kilograms}

\item{push_off_distance}{Numeric vector in meters. Default 0.4 meters}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector in m/s
}
\description{
\code{TOV_from_force} returns take off velocity when \code{force} is applied to the object
of \code{mass} over \code{push_off_distance} in vertical direction, assuming zero starting velocity
}
\examples{
TOV_from_force(
  force = 2000,
  mass = 85,
  push_off_distance = 0.4
)
}
