% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-funcs.R
\name{TOV_from_FT}
\alias{TOV_from_FT}
\title{Get Take-off Velocity from Flight Time}
\usage{
TOV_from_FT(flight_time, gravity_const = 9.81)
}
\arguments{
\item{flight_time}{Numeric vector. Flight time in seconds}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector. Take-off velocity in m/s
}
\description{
\code{height_from_FT} returns take-off velocity using \code{flight_time}
}
\examples{
TOV_from_FT(0.75)
}
