% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-funcs.R
\name{TOV_from_height}
\alias{TOV_from_height}
\title{Get Take-off Velocity from Height}
\usage{
TOV_from_height(height, gravity_const = 9.81)
}
\arguments{
\item{height}{Numeric vector in meters}

\item{gravity_const}{Numeric vector. Default 9.81}
}
\value{
Numeric vector. Flight time in seconds
}
\description{
\code{TOV_from_height} returns take-off velocity using \code{height}
}
\examples{
TOV_from_height(0.3)
}
